% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campsite.R
\name{cs_simulate_reps}
\alias{cs_simulate_reps}
\title{Run replicates of simulation using the Competition And Multiple-Peak Selection
Integrated Trait Evolution (CAMPSITE) model}
\usage{
cs_simulate_reps(
  pars,
  ou = list(opt = NULL, alpha4 = NULL),
  root.value = 0,
  age.max = 50,
  age.ext = NULL,
  step_size = 0.01,
  bounds = c(-Inf, Inf),
  plot = TRUE,
  ylims = NULL,
  full_results = TRUE,
  out_filepath = NULL,
  max_rep = 100,
  ensure_valid = FALSE
)
}
\arguments{
\item{pars}{an object of class \code{cs_pars}.}

\item{ou}{a named list of length 2 containing a vector of \code{optima} and \code{alpha4} values.}

\item{root.value}{numeric. The initial trait value of the root node. Must be within the specified bounds.}

\item{age.max}{numeric. maximum age of simulation}

\item{age.ext}{age of extinction}

\item{step_size}{size of each simulation step}

\item{bounds}{numeric vector of length 2. Trait value bounds}

\item{plot}{logical. Whether to plot resulting tree}

\item{ylims}{ylim}

\item{full_results}{logical. Whether to return full results, including lineage and}

\item{out_filepath}{file path to save simulation replicate output}

\item{max_rep}{maximum number of simulation replicates}

\item{ensure_valid}{whether to to repeat simulate until valid (process not dead and nNode > 4) replicate obtained}
}
\value{
The function is intended to write results out to disk so results are returned invisibly.
Output is a list with one element for each replicate. Each simulation result is an object of class \code{cs_sim_results}
}
\description{
Run replicates of simulation using the Competition And Multiple-Peak Selection
Integrated Trait Evolution (CAMPSITE) model
}
