% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{cs_summarise_results}
\alias{cs_summarise_results}
\title{Summarise results of CAMPSITE simulation}
\usage{
cs_summarise_results(x, max_trait_len = NULL)
}
\arguments{
\item{x}{results object of class \code{"cs_sim_results"} output from \code{cs_simulate()}
or list of \code{"cs_sim_results"} objects.}

\item{max_trait_len}{length each trait vector should be expanded by.}
}
\value{
an object (if \code{x} is \code{"cs_sim_results"} object) or list objects
(if \code{x} is list of \code{"cs_sim_results"} objects) of class \code{cs_result_summaries} containing:
\itemize{
\item \code{lineages}: object of class \code{cs_lineages}.
\item \code{VAR}: variance of trait values across all lineages at each time step
\item \code{MNND}: Mean Nearest Neighbour Distance of trait values across all lineages at each time step.
\item \code{VNND}: Variance of Nearest Neighbour Distance of trait values across all lineages at each time step.
\item \code{tip_traits}: tip trait values of extant lineages
\item \code{Nnode_extant}: extant tree number of nodes
\item \code{tree_extant}: object of class \code{phylo} containing extant lineages only
\item \code{tree_fossil}: object of class \code{phylo} containing tree that excludes incomplete lineages
}
}
\description{
Summarise results of CAMPSITE simulation
}
