% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_lineage}
\alias{new_lineage}
\title{Create a lineage object}
\usage{
new_lineage(x = numeric(length = 7L))
}
\arguments{
\item{x}{numeric vector of length 7.}
}
\value{
An object of class "lineage" with the following values
\itemize{
\item \code{parent_node} parental node (lineage number)`
\item \code{descendant_node} descendent node (0 if tip)
\item \code{start_time} starting time
\item \code{end_time} ending time (-1 if still active)
\item \code{status} status (incipient=-1/good=1/extinct=-2)
\item \code{spec_or_ext_ct} speciation completion or extinction time
\item \code{spec_ct} speciation completion time (NA if still incipient)
}
}
\description{
Create a lineage object
}
