% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cs_sim_results}
\alias{plot.cs_sim_results}
\title{Plot CAMPSITE simulation results}
\usage{
\method{plot}{cs_sim_results}(x, ylims = NULL)
}
\arguments{
\item{x}{results object of class "cs_sim_results" output from \code{sim_CAMPSITE}.}

\item{ylims}{numeric vector of length 2 specifying limits for the y axis. If
\code{NULL} (default), limits are determined automatically}
}
\value{
plot of simulated lineage evolution across trait space
}
\description{
Plot CAMPSITE simulation results
}
\examples{
\dontrun{
res <- cs_simulate(pars = cs_pars(), ou = list(opt = NULL, alpha4 = NULL), root.value = 0, age.max = 10, 
                   age.ext = NULL, step_size = 0.01, bounds = c(-Inf, Inf), 
                   plot = FALSE, ylims = NULL, full_results = TRUE) 

plot(res)                   
}

}
