% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_diversification}
\alias{plot_diversification}
\title{Plot result diversification rates against time.}
\usage{
plot_diversification(x)
}
\arguments{
\item{x}{an object or list of objects of class \code{cs_result_summaries}.}
}
\value{
plot of diversification against time
}
\description{
Plot result diversification rates against time.
}
