% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_diversification_replicates}
\alias{plot_diversification_replicates}
\title{Plot replicate diversification rates against time for a single competition x selection combination}
\usage{
plot_diversification_replicates(x)
}
\arguments{
\item{x}{an object or list of objects of class \code{cs_result_summaries}.}
}
\value{
plot of replicate diversification rates against time
}
\description{
Plot replicate diversification rates against time for a single competition x selection combination
}
