% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_lineages_through_time_replicates}
\alias{plot_lineages_through_time_replicates}
\title{Species richness across replicates for a single competition x selection combination}
\usage{
plot_lineages_through_time_replicates(x)
}
\arguments{
\item{x}{an object or list of objects of class \code{cs_result_summaries}. Must have replicate details recorded.}
}
\value{
a plot of species richness through time
}
\description{
Species richness across replicates for a single competition x selection combination
}
