% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_tip_trait_distribution}
\alias{plot_tip_trait_distribution}
\title{Plot tip trait distribution on competition x selection grid}
\usage{
plot_tip_trait_distribution(x)
}
\arguments{
\item{x}{an object of class \code{cs_result_summaries} or list of \code{cs_result_summaries} objects}
}
\value{
a plot of tip trait distributions
}
\description{
Plot tip trait distribution on competition x selection grid
}
