% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_tip_trait_distribution_replicates}
\alias{plot_tip_trait_distribution_replicates}
\title{Plot tip trait distribution across replicates for a single competition x selection combination}
\usage{
plot_tip_trait_distribution_replicates(x)
}
\arguments{
\item{x}{an object or list of objects of class \code{cs_result_summaries}. Must have replicate details recorded.}
}
\value{
a plot of tip trait distributions
}
\description{
Plot tip trait distribution across replicates for a single competition x selection combination
}
