% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_var_vs_time_replicates}
\alias{plot_var_vs_time_replicates}
\title{Plot result summary variable replicates for a single competition x selection combination against time.}
\usage{
plot_var_vs_time_replicates(
  x,
  variable = c("VAR", "MNND", "VNND"),
  ylab = NULL,
  time = c("step", "int")
)
}
\arguments{
\item{x}{an object or list of objects of class \code{cs_result_summaries}. Must have replicate details recorded.}

\item{variable}{character string. The variable to be plotted through time.}

\item{ylab}{character string. Label for y axis. Defaults to variable.}

\item{time}{character string. If \code{step}, values are plotted against each time step.
If \code{int} the  values at each integer time point are plotted}
}
\value{
plot of variable mean and replicates for a single competition x selection combination against time. The last replicate is highlighted.
}
\description{
Plot result summary variable replicates for a single competition x selection combination against time.
}
