% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_trait}
\alias{new_trait}
\title{Create a trait object}
\usage{
new_trait(
  lin_id = integer(1),
  status = integer(1),
  parent_id = integer(1),
  parent_trait_value = numeric(1),
  trait_values = numeric()
)
}
\arguments{
\item{lin_id}{lineage identifying number}

\item{status}{status (incipient=-1/good=1/extinct=-2)}

\item{parent_id}{parental node (lineage number)}

\item{trait_values}{current trait value of parental lineage, and all consecutive
elements contain trait value at each time step (always \code{root.value} at the beginning)}

\item{x}{a vector of four or more values.}
}
\value{
An object of class "trait" with the following values
}
\description{
Create a trait object
}
\examples{
new_trait(1, 1, 2, c(NA, root.value))
}
